Fix a compiler warning about comparing int to size_t which was functionally not relevant (the int was guaranteed by program logic to be >= 0), but which annoyed this commit author :)

+++ src/vm.c
@@ -491,7 +491,7 @@ static ms_VMValue *VMPeek(ms_VM *vm, int index) {
         assert(((int)f->dp + index) >= 0);
         return &f->data[f->dp + index];
     } else {
-        assert(index < f->dp);
+        assert((size_t)index < f->dp);
         return &f->data[index];
     }
 }
